rename libraries file
